add derive macros for StateRead and StateWrite
