+++ crates/core/component/ibc/src/component/rpc/client_query.rs
@@ -1,5 +1,7 @@
+use anyhow::Result;
 use async_trait::async_trait;
-
+use cnidarium::StateRead;
+use cnidarium_component::ChainStateReadExt;
 use ibc_proto::ibc::core::client::v1::query_server::Query as ClientQuery;
 use ibc_proto::ibc::core::client::v1::{
     Height, IdentifiedClientState, QueryClientParamsRequest, QueryClientParamsResponse,
@@ -11,8 +13,6 @@ use ibc_proto::ibc::core::client::v1::{
     QueryUpgradedClientStateResponse, QueryUpgradedConsensusStateRequest,
     QueryUpgradedConsensusStateResponse,
 };
-
-use cnidarium_component::ChainStateReadExt;
 use ibc_types::core::client::ClientId;
 use ibc_types::lightclients::tendermint::client_state::TENDERMINT_CLIENT_STATE_TYPE_URL;
 use ibc_types::path::ClientStatePath;
@@ -27,58 +27,11 @@ use crate::IBC_COMMITMENT_PREFIX;
 
 use super::IbcQuery;
 
-use cnidarium::StateRead;
-
+#[derive(wrapper_derive::StateRead)]
 struct SnapshotWrapper<S: StateRead>(S);
 
-impl<S: StateRead> StateRead for SnapshotWrapper<S> {
-    type GetRawFut = S::GetRawFut;
-    type PrefixRawStream = S::PrefixRawStream;
-    type PrefixKeysStream = S::PrefixKeysStream;
-    type NonconsensusPrefixRawStream = S::NonconsensusPrefixRawStream;
-    type NonconsensusRangeRawStream = S::NonconsensusRangeRawStream;
-
-    fn get_raw(&self, key: &str) -> Self::GetRawFut {
-        self.0.get_raw(key)
-    }
-
-    fn prefix_raw(&self, prefix: &str) -> S::PrefixRawStream {
-        self.0.prefix_raw(prefix)
-    }
-
-    fn prefix_keys(&self, prefix: &str) -> S::PrefixKeysStream {
-        self.0.prefix_keys(prefix)
-    }
-
-    fn nonverifiable_prefix_raw(&self, prefix: &[u8]) -> S::NonconsensusPrefixRawStream {
-        self.0.nonverifiable_prefix_raw(prefix)
-    }
-
-    fn nonverifiable_range_raw(
-        &self,
-        prefix: Option<&[u8]>,
-        range: impl std::ops::RangeBounds<Vec<u8>>,
-    ) -> anyhow::Result<Self::NonconsensusRangeRawStream> {
-        self.0.nonverifiable_range_raw(prefix, range)
-    }
-
-    fn nonverifiable_get_raw(&self, key: &[u8]) -> Self::GetRawFut {
-        self.0.nonverifiable_get_raw(key)
-    }
-
-    fn object_get<T: std::any::Any + Send + Sync + Clone>(&self, key: &'static str) -> Option<T> {
-        self.0.object_get(key)
-    }
-
-    fn object_type(&self, key: &'static str) -> Option<std::any::TypeId> {
-        self.0.object_type(key)
-    }
-}
-
-use anyhow::Result;
-
 #[async_trait]
-impl<S: cnidarium::StateRead + 'static> ChainStateReadExt for SnapshotWrapper<S> {
+impl<S: StateRead + 'static> ChainStateReadExt for SnapshotWrapper<S> {
     async fn get_chain_id(&self) -> Result<String> {
         use penumbra_chain::component::StateReadExt as _;
         self.0.get_chain_id().await
