+++ crates/util/wrapper-derive/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "wrapper-derive"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "1.0"
+quote = "1.0"
+proc-macro2 = "1.0"
